+++ crates/pms-utils/src/block_id.rs
@@ -50,6 +50,7 @@ pub fn compute_block_id(
                 PlainPayload::Reward { .. } => "Reward",
                 PlainPayload::EncryptedReward { .. } => "EncryptedReward",
                 PlainPayload::TokenCreate(_) => "TokenCreate",
+                PlainPayload::SftClassCreate(_) => "SftClassCreate",
                 PlainPayload::BridgeLock { .. } => "BridgeLock",
                 PlainPayload::BridgeMint { .. } => "BridgeMint",
                 PlainPayload::Freeze { .. } => "Freeze",